ms-common-ext/ms-common-ext-spring/ms-common-ext-spring-reactive/ms-common-ext-spring-reactive-filter/src/main/java/com/jrmcdonald/common/ext/spring/reactive/filter/config/ReactiveFilterConfiguration.java
package com.jrmcdonald.common.ext.spring.reactive.filter.config;

import com.jrmcdonald.common.ext.spring.reactive.filter.logging.ReactiveRequestLoggingFilter;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;

@Configuration
public class ReactiveFilterConfiguration {

    @Bean
    public ReactiveRequestLoggingFilter reactiveRequestLoggingFilter(@Value("${spring.application.name}") String applicationName) {
        return new ReactiveRequestLoggingFilter(applicationName);
    }
}
